finish moving over basic puzzle/solution funcs, time for the hard part
